Move `InlineAutocomplete` to `drafts`
