+++ src/drafts/InlineAutocomplete/InlineAutocomplete.stories.tsx
@@ -1,7 +1,7 @@
 import React, {useState} from 'react'
 import {Meta} from '@storybook/react'
 
-import {BaseStyles, Box, Textarea, ThemeProvider, FormControl} from '..'
+import {BaseStyles, Box, Textarea, ThemeProvider, FormControl} from '../..'
 import InlineAutocomplete, {ShowSuggestionsEvent, Suggestions} from '.'
 
 export default {
