+++ src/drafts/InlineAutocomplete/InlineAutocomplete.test.tsx
@@ -2,11 +2,11 @@ import React, {useState} from 'react'
 import {fireEvent, render, within} from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import InlineAutocomplete, {ShowSuggestionsEvent, Suggestions, Trigger} from '.'
-import FormControl from '../FormControl'
-import {ActionList} from '../ActionList'
-import Textarea from '../Textarea'
-import ThemeProvider from '../ThemeProvider'
-import useIsomorphicLayoutEffect from '../utils/useIsomorphicLayoutEffect'
+import FormControl from '../../FormControl'
+import {ActionList} from '../../ActionList'
+import Textarea from '../../Textarea'
+import ThemeProvider from '../../ThemeProvider'
+import useIsomorphicLayoutEffect from '../../utils/useIsomorphicLayoutEffect'
 
 const label = 'Inline Autocomplete'
