+++ src/drafts/InlineAutocomplete/types.ts
@@ -1,4 +1,4 @@
-import {ActionListItemProps} from '../ActionList'
+import {ActionListItemProps} from '../../ActionList'
 
 export type Trigger = {
   /** A single character that can cause the suggestion list to open. */
