+++ src/drafts/MarkdownEditor/suggestions/index.ts
@@ -1,4 +1,4 @@
-import {Suggestion, Trigger} from '../../../InlineAutocomplete'
+import {Suggestion, Trigger} from '../../InlineAutocomplete'
 
 const MAX_SUGGESTIONS = 5
