+++ src/drafts/index.ts
@@ -27,14 +27,14 @@ export type {
 
 export * from '../Dialog/Dialog'
 
-export {default as InlineAutocomplete} from '../InlineAutocomplete'
+export {default as InlineAutocomplete} from './InlineAutocomplete'
 export type {
   InlineAutocompleteProps,
   ShowSuggestionsEvent,
   Suggestion,
   Suggestions,
   Trigger,
-} from '../InlineAutocomplete'
+} from './InlineAutocomplete'
 
 export {default as MarkdownViewer} from './MarkdownViewer'
 export type {MarkdownViewerProps, InteractiveMarkdownViewerProps} from './MarkdownViewer'
